update phase 1 tests move method to ValidMoveTests
