+++ shared/src/test/java/passoff/chess/TestUtilities.java
@@ -69,12 +69,4 @@ public class TestUtilities {
         }
         return validMoves;
     }
-
-    public static void assertMoves(ChessGame game, Set<ChessMove> validMoves, ChessPosition position) {
-        var generatedMoves = game.validMoves(position);
-        var actualMoves = new HashSet<>(generatedMoves);
-        Assertions.assertEquals(generatedMoves.size(), actualMoves.size(), "Duplicate move");
-        Assertions.assertEquals(validMoves, actualMoves,
-                "ChessGame validMoves did not return the correct moves");
-    }
 }
